refactor eqV2 heads
