+++ src/fairchem/core/models/equiformer_v2/heads/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .rank2 import Rank2SymmetricTensorHead
+from .scalar import EqV2ScalarHead
+from .vector import EqV2VectorHead
+
+__all__ = ["EqV2ScalarHead", "EqV2VectorHead", "Rank2SymmetricTensorHead"]
