--- src/fairchem/core/models/equiformer_v2/prediction_heads/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from .rank2 import Rank2SymmetricTensorHead
-
-__all__ = ["Rank2SymmetricTensorHead"]
